requirements.txt
torch==2.6.0
torchvision==0.21.0
#torch==2.6.0+cu118
#torchvision==0.21.0+cu118
ultralytics==8.3.96
opencv-python>=4.5.1
numpy>=1.20.0
matplotlib==3.10.1
PyYAML==6.0
tqdm==4.64.0
scikit-learn==1.6.1
ipywidgets==8.1.5
natsort
shapely
pandas==2.2.2
Pillow==10.2.0